fix(use-promisify): add default args at weapp-fix
