+++ packages/use-promisify/test/index.test.js
@@ -1,5 +1,5 @@
+import usePromisifyInstall from '../install';
 const expect = require('chai').expect;
-const usePromisify = require('../dist/index');
 
 function ensureAllTaskDone(taskList, done) {
   let tasks = {};
@@ -72,7 +72,7 @@ describe('wepy-use-promisify', function() {
     let task = ensureAllTaskDone(['test-request-catch', 'test-storage'], done);
 
     let wepy = {};
-    usePromisify.install(wepy);
+    usePromisifyInstall(wepy);
 
     expect(wepy.wx.request).is.a('function');
 
@@ -97,7 +97,7 @@ describe('wepy-use-promisify', function() {
   it('install appends array list', function(done) {
     let wepy = {};
 
-    usePromisify.install(wepy, { someNewAPI: false, getStorage: true });
+    usePromisifyInstall(wepy, { someNewAPI: false, getStorage: true });
 
     wepy.wx.someNewAPI({ num: 1 }).then(res => {
       expect(res).to.equal(1);
@@ -107,7 +107,7 @@ describe('wepy-use-promisify', function() {
 
   it('install get rid apis', function(done) {
     let wepy = {};
-    usePromisify.install(wepy, ['getStorage']);
+    usePromisifyInstall(wepy, ['getStorage']);
     wepy.wx.getStorage({
       key: 'mydata',
       success: function(res) {
@@ -119,7 +119,7 @@ describe('wepy-use-promisify', function() {
 
   it('params fix testing', function(done) {
     let wepy = {};
-    usePromisify.install(wepy);
+    usePromisifyInstall(wepy);
 
     wepy.wx.setStorage('mydata', { b: 1 }).then(() => {
       wepy.wx.getStorage('mydata').then(res => {
@@ -144,7 +144,7 @@ describe('wepy-use-promisify', function() {
     };
 
     let wepy = {};
-    usePromisify.install(wepy);
+    usePromisifyInstall(wepy);
 
     let promisifyFn = wepy.promisify(isGreaterThan10, null, 'error-first');
 
@@ -176,7 +176,7 @@ describe('wepy-use-promisify', function() {
     };
 
     let wepy = {};
-    usePromisify.install(wepy);
+    usePromisifyInstall(wepy);
 
     let promisifyFn = wepy.promisify(isGreaterThan10);
 
@@ -218,7 +218,7 @@ describe('wepy-use-promisify', function() {
         done
       );
 
-    usePromisify.install(wepy);
+    usePromisifyInstall(wepy);
 
     for (let item of originParamsArr) {
       let filterParams = Object.values(item);
